pyproject.toml
[tool.poetry]
name = "smm-nq-trader"
version = "0.1.0"
description = "Ultra-low latency auto-trading platform for NQ futures"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
uvloop = "^0.20.0"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
websockets = "^12.0"
pyyaml = "^6.0.2"
pandas = "^2.2.2"
numpy = "^2.1.1"
async-rithmic = "*"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
